requirements.txt
scikit-learn==1.7.2
joblib==1.4.2
numpy
pandas
fastapi
uvicorn[standard]
google-cloud-firestore
requests
